Add new @cifailonly VK tag
For tests that pass locally but fail on the CI, this adds a new
test tag. These tests should continue to be tagged @xfail for now
as this is the only excluded tag in the CI. However this new tag
provides a simpler way to identify tests that are causing issues
in a CI job.